// mobile/gen/models/FindAppointments.ts
export const dataStatusEnum = {
  AGENDADA: "AGENDADA",
  CANCELADA: "CANCELADA",
  REAGENDADA: "REAGENDADA",
  CONCLUIDA: "CONCLUIDA",
} as const;

export type DataStatusEnumKey =
  (typeof dataStatusEnum)[keyof typeof dataStatusEnum];

export const dataPriorityEnum = {
  BAIXA: "BAIXA",
  MEDIA: "MEDIA",
  ALTA: "ALTA",
} as const;

export type DataPriorityEnumKey =
  (typeof dataPriorityEnum)[keyof typeof dataPriorityEnum];

export const dataSourceEnum = {
  TRIAGEM: "TRIAGEM",
  DIRECTA: "DIRECTA",
} as const;

export type DataSourceEnumKey =
  (typeof dataSourceEnum)[keyof typeof dataSourceEnum];